feat: Add metadata fin, actions, vieprivee
