+++ src/types/fin.ts
@@ -1,0 +1,1 @@
+export type Diapo = 'bilan' | 'categories' | 'actions'
